pyproject.toml
[tool.poetry]
name = "hotel-booking"
version = "0.1.0"
description = "Hotel Room and Booking Management System"
authors = ["Kirill <soldatovk1572@gmail.com>"]
readme = "README.md"
packages = [{include = "hotel_booking", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.7"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.9"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
ruff = "^0.13.3"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "N",    # pep8-naming
]
ignore = [
    "E203",  # Whitespace before ':'
]

[tool.ruff.lint.isort]
known-first-party = ["hotel_booking", "api"]
known-third-party = ["django", "rest_framework"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "hotel_booking.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["src", "tests"]
pythonpath = ["src"]
